backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg[binary]
python-dotenv
dj-database-url
psycopg
djangorestframework-simplejwt
celery
requests
redis
boto3